Teacher app: attendance audit remediation + regularization/leave/events/gallery bundle
Enterprise attendance health-audit fixes (verified, compiles clean):
- Student marking: long-press "mark Late for today" -> arrival-time dialog feeds
  late[]/arrivalTime (closes the parent lateTimes gap that had no current-day path).
- My Attendance (GPS): clock-in hint is advisory only (button gated on location
  permission, server stays sole authority) so a wrong device clock can't lock out
  a legitimate punch; calendar day-cell status semantics + 48dp touch targets;
  window-aware schedule hint (StaffScheduleHint).
- Regularization: deterministic docId {school}_{staff}_{date} + merge so a
  double-submit can't create duplicate pending rows; full-screen RegularizeSheet
  + My Requests, appliedMark surfacing.
- Leave: Firestore-first submit/cancel/history with typed states; robustness pass.

Plus previously-uncommitted cross-module work:
- Events: EventCover + screen/VM polish.
- Gallery: doc + repository fixes.
- Dashboard: schedule/agenda + error-state tweaks.
- DI (AppModule), NotificationChannels per-category channels.

+++ app/src/main/java/com/schoolsync/teacher/data/model/firestore/LeaveApplicationDoc.kt
@@ -15,7 +15,7 @@ data class LeaveApplicationDoc(
     val leaveType: String = "",      // "SL", "CL", "EL", "ML", "LWP"
     val startDate: String = "",
     val endDate: String = "",
-    val numberOfDays: Int = 0,
+    val numberOfDays: Double = 0.0,   // MEDIUM #5: Double so half-day = 0.5
     val reason: String = "",
     val attachments: List<String> = emptyList(),
     val status: String = "pending",  // "pending", "approved", "rejected", "cancelled"
